Safari-609.1.11.1/LayoutTests/fast/css/CSSStyleDeclaration-cssFloat-expected.txt

Test support for CSSStyleDeclaration.cssFloat

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'cssFloat' in testDiv.style is true
PASS Object.getOwnPropertyNames(testDiv.style).indexOf('cssFloat') is -1
PASS Object.getOwnPropertyNames(Object.getPrototypeOf(testDiv.style)).indexOf('cssFloat') != -1 is true
PASS testDiv.style.cssFloat is ""
testDiv.style.cssFloat = 'right'
PASS testDiv.style.cssFloat is "right"
testDiv.style.cssFloat = 'invalid'
PASS testDiv.style.cssFloat is "right"
testDiv.style.cssFloat = 'left'
PASS testDiv.style.cssFloat is "left"
testDiv.style.cssFloat = null
PASS testDiv.style.cssFloat is ""
PASS successfullyParsed is true

TEST COMPLETE
